Att report

+++ src/main/java/app/comunication/response/JwtResponse.java
@@ -4,12 +4,12 @@ package app.comunication.response;
 public class JwtResponse {
   private String token;
   private String type = "Bearer";
-  private int id;
+  private long id;
   private String username;
   private String role;
 
 
-  public JwtResponse(String accessToken, int id, String username, String role) {
+  public JwtResponse(String accessToken, long id, String username, String role) {
     this.token = accessToken;
     this.id = id;
     this.username = username;
@@ -32,7 +32,7 @@ public class JwtResponse {
     this.type = tokenType;
   }
 
-  public int getId() {
+  public long getId() {
     return id;
   }
